pyproject.toml
[project]
name = "megara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "fastexcel>=0.18.0",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
]
